Buildings menu and placing, energy, MSGUI -> MGUI
A building menu now opens when selecting a tile in-game, displaying placeable buildings owned by the player. Buildings may be placed in the owned area of the game if you can afford it.
In-game currency "energy" added, represented as a purple flame.
Renamed MSGUI (Mosca's Simple GUI) to MGUI due to the lack of claimed simplicity.