Timed effect buildings, level map layers, energy per second
player_owned variable for buildings, to distinguish from enemy buildings.
New building class BuildingTimedEffect, for buildings that produce an effect periodically. BuildingSpawner now a subclass of BuildingTimedEffect.
LevelMap now has a tiles and buildings layer, which are blitted to the main level_surface when an update is needed.
Energy per second gaining.