Multiple animations and animation sequencing for animsprites
Animated sprites now have a dictionary holding each possible animation. The default animation is "base". Each key in this dictionary contains a list of ints with the frames for the animation in order. These lists can contain a string, which will switch the animation at that point to the one named in the string.
Units can now use "idle", "base", "battle" and "death" as base animations for each state ("base" is required).